refactor: modularize Form
